requirements.txt
decorator==5.1.1
mysql-connector-python==8.0.29
protobuf==3.20.1
PyJWT==2.6.0
validators==0.20.0
